+++ divio_cli/api_requests.py
@@ -419,3 +419,26 @@ class ApplicationRequest(JsonResponse, APIV3Request):
 class CreateApplicationRequest(JsonResponse, APIV3Request):
     url = "/apps/v3/applications/create_with_template/"
     method = "POST"
+class ListServiceInstancesRequest(JsonResponse, APIV3Request):
+    url = "/apps/v3/service-instances/?environment={environment_uuid}"
+    method = "GET"
+
+
+class CreateServiceInstanceRequest(JsonResponse, APIV3Request):
+    url = "/apps/v3/service-instances/"
+    method = "POST"
+
+
+class ListServicesRequest(JsonResponse, APIV3Request):
+    url = "/apps/v3/services/?{filter_region}&{filter_website}"
+    method = "GET"
+
+
+class ListRegionsRequest(JsonResponse, APIV3Request):
+    url = "/apps/v3/regions/"
+    method = "GET"
+
+
+class ListOrganisationsRequest(JsonResponse, APIV3Request):
+    url = "/iam/v3/organisations/"
+    method = "GET"